\name{timss2007}
\alias{timss2007}
\docType{data}
\title{
Science Perception of Eigth Grade Students
}
\description{
This dataset gives the perception of science and attitude towards science for eigth grade students. Data is obtained from TIMSS 2007 Student Questionnaire. 
}
\usage{data(timss2007)}
\format{
  It consists of 4498 students with their 28 perception variables. Variables have been assigned according to last four letters of their coding in the Questionnaire 
  \describe{
    \item{\code{AWEL}}{ BS4SAWEL - I usually do well in science}
    \item{\code{AMOR}}{ BS4SAMOR - I would like to take more science in school}
    \item{\code{ACLM}}{ BS4SACLM - Science is more difficult for me than for many of my classmates}
    \item{\code{AENJ}}{ BS4SAENJ - I enjoy learning Science}
    \item{\code{ASTR}}{ BS4SASTR - Science is not one of my strengths}
    \item{\code{AQKY}}{ BS4SAQKY - I learn things quickly in Science}
    \item{\code{ABOR}}{ BS4SABOR - Science is boring}
    \item{\code{ALIK}}{ BS4SALIK - I like Science}
    \item{\code{AHDL}}{ BS4SAHDL - I think learning Science will help me in my daily life}
    \item{\code{AOSS}}{ BS4SAOSS - I need Science to learn other school subjects}
    \item{\code{AUNI}}{ BS4SAUNI - I need to do well in Science to get into the university of my choice}
    \item{\code{AGET}}{ BS4SAGET - I need to do well in Science to get the job I want}
    \item{\code{HOBS}}{ BS4SHOBS - In your science lessons, how often do you make observations and describe what you see?}
    \item{\code{HDEI}}{ BS4SHDEI - In your science lessons, how often do you watch the teacher demonstrate an experiment or investigation?}
    \item{\code{HPEI}}{ BS4SHPEI - In your science lessons, how often do you design or plan an experiment or investigation?}
    \item{\code{HCEI}}{ BS4SHCEI - In your science lessons, how often do you conduct an experiment or investigation?}
    \item{\code{HWGO}}{ BS4SHWGO - In your science lessons, how often do you work in small groups on an experiment or investigation?}
    \item{\code{HTEX}}{ BS4SHTEX - In your science lessons, how often do you read your science text books and other resource materials?}
    \item{\code{HFAP}}{ BS4SHFAP - In your science lessons, how often do you memorize science facts and principles?}
    \item{\code{HLAW}}{ BS4SHLAW - In your science lessons, how often do you use scientific formulas and laws to solve problems?}
    \item{\code{HEOS}}{ BS4SHEOS - In your science lessons, how often do you give explanations about what you are studying?}
    \item{\code{HMDL}}{ BS4SHMDL - In your science lessons, how often do you relate what you are learning in science to your daily lives?}
    \item{\code{HROH}}{ BS4SHROH - In your science lessons, how often do you review your homework?}
    \item{\code{HLSP}}{ BS4SHLSP - In your science lessons, how often do you listen to the teacher give a lecture-style presentation?}
    \item{\code{HWPO}}{ BS4SHWPO - In your science lessons, how often do you work problems on your own?}
    \item{\code{HBHC}}{ BS4SHBHC - In your science lessons, how often do you begin your homework in class?}
    \item{\code{HHQT}}{ BS4SHHQT - In your science lessons, how often do you have a quiz or test?}
    \item{\code{HCOM}}{ BS4SHCOM - In your science lessons, how often do you use computers?}
  }
}
\source{
TIMSS 2007 Student Questionnaire http://timss.bc.edu/timss2007/idb_ug.html
}
\references{
TIMSS 2007 User Guide Supplement 1 - International Version of the TIMSS 2007 Background and Curriculum Questionnaires, 2009, Page 166
}
\keyword{datasets}
