\name{timss2011}
\alias{timss2011}
\docType{data}
\title{
Science Perception of Eigth Grade Students
}
\description{
This dataset gives the perception of science and attitude towards science for eigth grade students in Turkey data of TIMSS 2011. Data is obtained from TIMSS 2011 Student Questionnaire. Among 42 questions, first 16 questions are on students' general opinion about school, and last 26 questions are about students' attitude and perception towards science lessons. All the variables are measured with likert type scale with four choices where '1' indicates "Every Day or Almost Every Day / Agree A Lot / At Least Once A Week" and '4' indicates "Never or Almost Never / Disagree A Lot / Never" depending on the question. 
}
\usage{data(timss2011)}
\format{
  Data consists of 6925 students with their 42 perception variables. Variables have been assigned according to last three letters of their coding in the questionnaire.
  \describe{
    \item{\code{X10A}}{ How often do you use a computer at home?}
    \item{\code{X10B}}{ How often do you use a computer at school?}
    \item{\code{X10C}}{ How often do you use a computer at some other place?}
    \item{\code{X11A}}{ How often do your parents ask what you learned in school?}
    \item{\code{X11B}}{ How often do you talk about schoolwork with your parents at home?}
    \item{\code{X11C}}{ How often do your parents make sure that you set aside time for your homework?}
    \item{\code{X11D}}{ How often do your parents check if you do your homework?}
    \item{\code{X12A}}{ How much do you agree that you like being in school? }
    \item{\code{X12B}}{ How much do you agree that you feel safe when you are at school?}
    \item{\code{X12C}}{ How much do you agree that you feel like you belong at your school?}
    \item{\code{X13A}}{ During this year, how often were you made fun of or called names at school?}
    \item{\code{X13B}}{ During this year, how often were you left out of games or activities by other students at school?}
    \item{\code{X13C}}{ During this year, how often did someone spread lies about you at school?}
    \item{\code{X13D}}{ During this year, how often was something stolen from you at school?}
    \item{\code{X13E}}{ During this year, how often were you hit or hurt by other student(s) at school?}
    \item{\code{X13F}}{ During this year, how often were you made to do things you didn't want to do by other students at school?}
    \item{\code{X17A}}{ How much do you agree that you enjoy learning science? }
    \item{\code{X17B}}{ How much do you agree that you wish you did not have to study science?}
    \item{\code{X17C}}{ How much do you agree that you read about science in your spare time?}
    \item{\code{X17D}}{ How much do you agree that science is boring?}
    \item{\code{X17E}}{ How much do you agree that you learn many interesting things in science?}
    \item{\code{X17F}}{ How much do you agree that you like science? }
    \item{\code{X17G}}{ How much do you agree that it is important to do well in science?}
    \item{\code{X18A}}{ How much do you agree that you know what your teacher expects you to do in your science lessons?}
    \item{\code{X18B}}{ How much do you agree that you think of things not related to the lesson in your science lessons?}
    \item{\code{X18C}}{ How much do you agree that your teacher is easy to understand in your science lessons?}
    \item{\code{X18D}}{ How much do you agree that you are interested in what your teacher is saying in your science lessons?}
    \item{\code{X18E}}{ How much do you agree that your teacher gives you interesting things to do in your science lessons?}
    \item{\code{X19A}}{ How much do you agree that you usually do well in science?}
    \item{\code{X19B}}{ How much do you agree that science is more difficult for you than for many of your classmates?}
    \item{\code{X19C}}{ How much do you agree that science is not one of your strengths?}
    \item{\code{X19D}}{ How much do you agree that you learn things quickly in science?}
    \item{\code{X19E}}{ How much do you agree that science makes you confused and nervous?}
    \item{\code{X19F}}{ How much do you agree that you are good at working out difficult science problems?}
    \item{\code{X19G}}{ How much do you agree that your teacher thinks you can do well in science <programs/classes/lessons> with difficult materials?}
    \item{\code{X19H}}{ How much do you agree that your teacher tells you that you are good at science?}
    \item{\code{X19I}}{ How much do you agree that science is harder for you than any other subject?}
    \item{\code{X19J}}{ How much do you agree that learning science will help you in your daily life?}
    \item{\code{X19K}}{ How much do you agree that you need science to learn other school subjects?}
    \item{\code{X19L}}{ How much do you agree that you need to do well in science to get into the <university> of your choice?}
    \item{\code{X19M}}{ How much do you agree that you need to do well in science to get the job you want?}
    \item{\code{X19N}}{ How much do you agree that you would like a job that involves using science?}
  }
}
\source{
TIMSS 2011 Student Questionnaire http://timssandpirls.bc.edu/timss2011/frameworks.html
}
\references{
TIMSS 2011 User Guide for the International Database. 2013 International Association for the Evaluation of Educational Achievement 
}
\keyword{TIMSS}
