\name{wqporsuk}
\alias{wqporsuk}
\docType{data}
\title{
Water Quality Data of Porsuk Stream
}
\description{
This correlation matrix has been obtained from the surface water quality data of Porsuk stream, Kutahya, Turkey. The matrix contains correlation of variables collected in winter season.   
}
\usage{data(wqporsuk)}
\format{
    29x29 correlation matrix has been generated with variables below
\describe{
    \item{\code{Q}}{Flow (\eqn{m^3}/\eqn{sn})}
    \item{\code{T}}{Temperature (\eqn{C})}
    \item{\code{pH}}{pH}
    \item{\code{EC}}{Electrical Conductivity (\eqn{umho}/\eqn{cm})}
    \item{\code{TDS}}{Total Dissolved Solids (\eqn{mg}/\eqn{l})}
    \item{\code{SS}}{Suspended Solids (\eqn{mg}/\eqn{l})}
    \item{\code{M-Al}}{Total Alkalinity (\eqn{mg}/\eqn{l})}
    \item{\code{Cl-}}{Chloride (\eqn{mg}/\eqn{l})}
    \item{\code{NH3-N}}{Ammonical Nitrogen (\eqn{mg}/\eqn{l})}
    \item{\code{NO2-N}}{Nitrite Nitrogen (\eqn{mg}/\eqn{l})}
    \item{\code{NO3-N}}{Nitrate Nitrogen (\eqn{mg}/\eqn{l})}
    \item{\code{DO}}{Dissolved Oxygen (\eqn{mg}/\eqn{l})}
    \item{\code{pV}}{Organic Matter (\eqn{mg}/\eqn{l})}
    \item{\code{BOD5}}{Biochemical Oxygen Demand (\eqn{mg}/\eqn{l})}
    \item{\code{COD}}{Chemical Oxygen Demand (\eqn{mg}/\eqn{l})}
    \item{\code{TH}}{Total Hardness (\eqn{mg}/\eqn{l})}
    \item{\code{o-PO4}}{Orto Phosphorus (\eqn{mg}/\eqn{l})}
    \item{\code{SO4}}{Sulphate (\eqn{mg}/\eqn{l})}
    \item{\code{FE}}{Iron (\eqn{mg}/\eqn{l})}
    \item{\code{Mn}}{Manganez (\eqn{mg}/\eqn{l})}
    \item{\code{Na+}}{Sodium (\eqn{mg}/\eqn{l})}
    \item{\code{K+}}{Potasium (\eqn{mg}/\eqn{l})}
    \item{\code{Ca2+}}{Caldium (\eqn{mg}/\eqn{l})}
    \item{\code{Mg2+}}{Magnesium (\eqn{mg}/\eqn{l})}
    \item{\code{Tcoli}}{Total Coliforms (\eqn{MPN}/\eqn{100 ml})}
    \item{\code{Fstrp}}{Fecal Streptococ (\eqn{MPN}/\eqn{100 ml})}
    \item{\code{Ecoli}}{Escherichia coli (\eqn{MPN}/\eqn{100 ml})}
    \item{\code{B}}{Boron (\eqn{mg}/\eqn{l})}
    \item{\code{F-}}{Fluoride (\eqn{mg}/\eqn{l})}
}
}
\references{
Altin, A.  Filiz, Z. Iscen, C.F. (2008) Assessment of seasonal variations of surface water quality characteristics for Porsuk Stream. \emph{Environ Monit Assess} 
158:51-65
}
\keyword{datasets}
